Ensure the index is always stored alongside the Index
This is necessary if an Index occurs in a Cond branch, where its index
argument is not computed inside the branch. Before this change, the
saved index in the index adjoint tuple would be the one computed before
the branch -- which is not set to -1 if the branch is not taken. With
this change, the Index uses an index value that is re-bound inside the
Cond branch, ensuring that it is -1 in the index adjoint tuple if the
containing branch is not taken.