Fix DMap.fromList discarding indexing contribs into same tuple
This was found by trying to run fusion before AD. It seems that when
presented with code like the following:

  let T2 a b = ...
  in generate (...) (\i -> a ! ... + b ! ...)

in other words, an expression that indexes into two different arrays
bound in the same let-binding, sharing recovery adds an intermediate
let-binding:

  let T2 a b = ...
  in let c = a
         d = b
     in generate (...) (\i -> c ! ... + d ! ...)

For the first case, my code was incorrect because of fromList's footgun.
However, sharing recovery normally prevents me from triggering that
problem because of the additional let-binding.

Optimising before AD removed the blatantly useless let-binding, and
exposed the bug.

Problem: this is not testable. Due to the above behaviour by sharing
recovery, any attempt to create a situation like the first example with
user code will result in an additional intermediate let-binding, which
hides the problem. Oh well.